Исправлены ошибки в проверке доступа для рефакторинга и добавлена проверка доступа для удаления товара.

+++ routes/products.js
@@ -1,21 +1,26 @@
 const {Router} = require('express');
 const auth = require('../middleware/auth');
-const Products = require('../models/products')
+const Products = require('../models/products');
+const keys = require('../keys/index')
 
 const router = Router();
 
-function isAdmin(userId, req) {
-    return req.user._id.toString() === userId.toString();
+function isAdmin(id) {
+    return keys.ADMIN_ID.indexOf(id.toString()) >= 0;
 }
 
 router.get('/', async (req, res) => {
-    const productsList = await Products.find();
-    res.render('products', {
-        title: 'Товары',
-        isProducts: true,
-        userId: req.user ? req.user._id.toString() : null,
-        productsList
-    })
+    try {
+        const productsList = await Products.find();
+        res.render('products', {
+            title: 'Товары',
+            isProducts: true,
+            userId: req.user ? req.user._id.toString() : null,
+            productsList
+        })
+    } catch(err) {
+        console.log(err)
+    }
 })
 
 router.get('/:id/edit', auth, async (req, res) => {
@@ -26,7 +31,7 @@ router.get('/:id/edit', auth, async (req, res) => {
     
         const product = await Products.findById(req.params.id);
 
-        if(!isAdmin(product.userId, req)) {
+        if(!isAdmin(req.user._id)) {
            return res.redirect('/products')
         }
     
@@ -37,20 +42,28 @@ router.get('/:id/edit', auth, async (req, res) => {
     } catch(err) {
         console.log(err)
     }
-    
 })
 
 router.post('/edit', auth, async (req, res) => {
-    const {id} = req.body;
-    delete req.body.id;
-    await Products.findByIdAndUpdate(id, req.body);
+    try {
+        const {id} = req.body;
+        delete req.body.id;
 
-    res.redirect('/products')
+        if(isAdmin(req.user._id)) {
+            await Products.findByIdAndUpdate(id, req.body);
+        }
+
+        res.redirect('/products')
+    } catch(err) {
+        console.log(err)
+    }
 })
 
 router.post('/remove', auth, async (req, res) => {
     try {
-        await Products.deleteOne({_id: req.body.id});
+        if(isAdmin(req.user._id)) {
+            await Products.deleteOne({_id: req.body.id});
+        }
         res.redirect('/products')
     } catch(e) {
         console.log(e)
@@ -58,13 +71,17 @@ router.post('/remove', auth, async (req, res) => {
 })
 
 router.get('/:id', async (req, res) => {
-    const product = await Products.findById(req.params.id);
-    res.render('product', {
-        layout: 'empty',
-        section: 'products',
-        title: `Товар ${product.title}`,
-        product
-    })
+    try {
+        const product = await Products.findById(req.params.id);
+        res.render('product', {
+            layout: 'empty',
+            section: 'products',
+            title: `Товар ${product.title}`,
+            product
+        })
+    } catch(err) {
+        console.log(err)
+    }
 })
 
 module.exports = router
